feat: separate into multiple files for organization
